refactor: #567 年齢モード5種を [uiMode] パラメータルートに集約
5つの年齢モード別ディレクトリ (baby/preschool/elementary/junior/senior) を
SvelteKit パラメータマッチャー [uiMode=uiMode] による単一ルートに統合。

- 新規: src/params/uiMode.ts — UI_MODES + LEGACY_UI_MODE_MAP 対応パラメータマッチャー
- 新規: src/lib/features/child-home/variants/index.ts — モード別テキスト・フィーチャーフラグ設定
- 統合: home/+page.svelte — baby のインライン記録 vs 他モードの確認ダイアログをフィーチャーフラグで切替
- 統合: achievements, status, history — getModeVariant() でテキスト・表示制御
- 削除: 5モード × 4ページ = 20ファイル（約3,900行）→ 4ファイル（約1,100行）に圧縮

既存URL (/preschool/home 等) はパラメータマッチャーにより引き続き動作。
レガシーURL (/kinder/home 等) も LEGACY_UI_MODE_MAP 経由で動作。
