+++ src/routes/(child)/[uiMode=uiMode]/(character)/achievements/+page.server.ts
