+++ src/routes/(child)/[uiMode=uiMode]/(character)/history/+page.server.ts
