+++ src/routes/(child)/[uiMode=uiMode]/home/+page.server.ts
